phone-service/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
twilio>=8.5.0
deepgram-sdk>=2.11.0
openai>=1.3.0
requests>=2.31.0
termcolor>=2.3.0
websockets>=11.0.3
python-multipart>=0.0.6
